- renamed xxx.services -> xxx-services - changed some strings to EntityType - worked around EntityType type issue - sorted items in entityTypeCdToText()
