+++ src/mixins/nr-affiliation-mixin.ts
@@ -1,7 +1,7 @@
 import { Component, Mixins } from 'vue-property-decorator'
 import { Action } from 'vuex-class'
-import AuthServices from '@/services/auth.services'
-import BusinessServices from '@/services/business.services'
+import AuthServices from '@/services/auth-services'
+import BusinessServices from '@/services/business-services'
 import { BusinessRequest, NameRequestI } from '@/interfaces'
 import { ActionBindingIF } from '@/interfaces/store-interfaces'
 import { Navigate } from '@/plugins'
