+++ src/mixins/payment-mixin.ts
@@ -12,7 +12,7 @@ import errorModule from '@/modules/error'
 import { ErrorI } from '@/modules/error/store/actions'
 import { StaffPaymentIF, RefundParamsIF, NameRequestI } from '@/interfaces'
 import { ActionBindingIF } from '@/interfaces/store-interfaces'
-import NamexServices from '@/services/namex.services'
+import NamexServices from '@/services/namex-services'
 import { PaymentRequiredError } from '@/errors'
 import { Navigate } from '@/plugins'
 import { appBaseURL } from '../router/router'
